LPS-179214 Extract ScreenNavigationEntry methods from AccountGroupAccountEntriesScreenNavigationCategory to AccountGroupAccountEntriesScreenNavigationEntry, extends AccountGroupAccountEntriesScreenNavigationCategory and return getCategoryKey in getEntryKey method
