+++ modules/apps/account/account-admin-web/src/main/java/com/liferay/account/admin/web/internal/frontend/taglib/servlet/taglib/AccountGroupAccountEntriesScreenNavigationCategory.java
@@ -15,22 +15,10 @@
 package com.liferay.account.admin.web.internal.frontend.taglib.servlet.taglib;
 
 import com.liferay.account.admin.web.internal.constants.AccountScreenNavigationEntryConstants;
-import com.liferay.account.admin.web.internal.display.AccountGroupDisplay;
-import com.liferay.account.admin.web.internal.security.permission.resource.AccountGroupPermission;
-import com.liferay.account.constants.AccountActionKeys;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
 import com.liferay.portal.kernel.language.Language;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.security.permission.PermissionCheckerFactoryUtil;
-
-import java.io.IOException;
 
 import java.util.Locale;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -39,15 +27,11 @@ import org.osgi.service.component.annotations.Reference;
  * @author Albert Lee
  */
 @Component(
-	property = {
-		"screen.navigation.category.order:Integer=20",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.category.order:Integer=20",
+	service = ScreenNavigationCategory.class
 )
 public class AccountGroupAccountEntriesScreenNavigationCategory
-	implements ScreenNavigationCategory,
-			   ScreenNavigationEntry<AccountGroupDisplay> {
+	implements ScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
@@ -55,13 +39,8 @@ public class AccountGroupAccountEntriesScreenNavigationCategory
 	}
 
 	@Override
-	public String getEntryKey() {
-		return AccountScreenNavigationEntryConstants.ENTRY_KEY_ACCOUNTS;
-	}
-
-	@Override
 	public String getLabel(Locale locale) {
-		return _language.get(locale, "accounts");
+		return language.get(locale, "accounts");
 	}
 
 	@Override
@@ -70,35 +49,7 @@ public class AccountGroupAccountEntriesScreenNavigationCategory
 			SCREEN_NAVIGATION_KEY_ACCOUNT_GROUP;
 	}
 
-	@Override
-	public boolean isVisible(
-		User user, AccountGroupDisplay accountGroupDisplay) {
-
-		if (accountGroupDisplay.getAccountGroupId() == 0) {
-			return false;
-		}
-
-		return AccountGroupPermission.contains(
-			PermissionCheckerFactoryUtil.create(user),
-			accountGroupDisplay.getAccountGroupId(),
-			AccountActionKeys.VIEW_ACCOUNTS);
-	}
-
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		jspRenderer.renderJSP(
-			httpServletRequest, httpServletResponse,
-			"/account_groups_admin/account_group/view_account_entries.jsp");
-	}
-
 	@Reference
-	protected JSPRenderer jspRenderer;
-
-	@Reference
-	private Language _language;
+	protected Language language;
 
 }
